backend/requirements.txt
fastapi[standard]
uvicorn
sqlmodel
sqlalchemy
python-dotenv